Changes method naming, adda automapper
(cherry picked from commit ece984376fa341c93883147f36b2322ef1aa6c27)
