+++ packages/apps/human-app/server/src/app.module.ts
@@ -11,8 +11,10 @@ import { OperatorModule } from './modules/user-operator/operator.module';
 import { OperatorController } from './modules/user-operator/operator.controller';
 import { WorkerController } from './modules/user-worker/worker.controller';
 import { CommonConfigModule } from './common/config/common-config.module';
-import { JobsDiscoveryModule } from './modules/jobs-discovery/jobs-discovery.module';
-import { JobsDiscoveryController } from './modules/jobs-discovery/jobs-discovery.controller';
+import { CacheFactoryConfig } from './common/config/cache-factory.config';
+import { CacheModule } from '@nestjs/cache-manager';
+import { OracleDiscoveryController } from './modules/oracle-discovery/oracle-discovery.controller';
+import { OracleDiscoveryModule } from './modules/oracle-discovery/oracle-discovery.module';
 
 @Module({
   imports: [
@@ -24,18 +26,19 @@ import { JobsDiscoveryController } from './modules/jobs-discovery/jobs-discovery
     AutomapperModule.forRoot({
       strategyInitializer: classes(),
     }),
+    CacheModule.registerAsync(CacheFactoryConfig),
     HttpModule,
     WorkerModule,
     OperatorModule,
-    JobsDiscoveryModule,
     ReputationOracleModule,
     CommonConfigModule,
+    OracleDiscoveryModule,
   ],
   controllers: [
     AppController,
     OperatorController,
     WorkerController,
-    JobsDiscoveryController,
+    OracleDiscoveryController,
   ],
   providers: [],
   exports: [HttpModule],
