--- packages/apps/human-app/server/src/modules/job-discovery/interface/oracle-discovery.interface.ts
@@ -1,20 +1,0 @@
-import { ChainId, IOperator } from '@human-protocol/sdk';
-export class OracleDiscoveryDto {
-  chainId: ChainId;
-  address: string;
-  role: string;
-}
-export class OracleDiscoveryCommand {
-  chainId: ChainId;
-  address: string;
-  role: string;
-}
-
-export class OracleDiscoveryData implements IOperator {
-  address: string;
-  role?: string;
-  constructor(address: string, role: string) {
-    this.address = address;
-    this.role = role;
-  }
-}
