+++ packages/apps/human-app/server/src/modules/oracle-discovery/oracle-discovery.mapper.ts
@@ -3,12 +3,11 @@ import { AutomapperProfile, InjectMapper } from '@automapper/nestjs';
 import { createMap, Mapper } from '@automapper/core';
 import {
   OracleDiscoveryCommand,
-  OracleDiscoveryData,
   OracleDiscoveryDto,
 } from './interface/oracle-discovery.interface';
 
 @Injectable()
-export class JobDiscoveryProfile extends AutomapperProfile {
+export class OracleDiscoveryProfile extends AutomapperProfile {
   constructor(@InjectMapper() mapper: Mapper) {
     super(mapper);
   }
